requirements.txt
# Core Django
Django>=5.0,<5.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Environment & Configuration
django-environ>=0.11.2

# Authentication
djangorestframework-simplejwt>=5.3.1

# API Documentation
drf-spectacular>=0.27.0

# Image handling
Pillow>=10.2.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.0.0

# Code Quality
flake8>=7.0.0
black>=23.12.1
isort>=5.13.2

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Monitoring
django-health-check>=3.18.1
